+++ clang-tools-extra/clang-tidy/misrac/MisracTidyModule.cpp
@@ -10,6 +10,7 @@
 #include "../ClangTidyModuleRegistry.h"
 #include "Rule10p1Check.h"
 #include "Rule10p2Check.h"
+#include "Rule10p3Check.h"
 
 namespace clang::tidy {
 namespace misrac {
@@ -21,6 +22,8 @@ public:
         "misrac-rule-10p1");
     CheckFactories.registerCheck<Rule10p2Check>(
         "misrac-rule-10p2");
+    CheckFactories.registerCheck<Rule10p3Check>(
+        "misrac-rule-10p3");
   }
 };